+++ src/components/Chat/index.ts
@@ -1,2 +1,3 @@
 export { default } from './Chat'
+export { default as ChatItem } from './ChatItem'
 export { default as ChatMessage, IChatMessageProps } from './ChatMessage'